General: converted remaining modules to double precision Physics: optimized physics a lot, both broad and narrow phases Rendering: - picking pass merged with geometry pass - tiling was added in geometry pass Scene: - lights and all other object types share now one scene graph so everything is updated with one function call - separate render lists for different object types - all objects are now indexed - only geometry objects are added to the picking list LineNode: LineNodes have now local positions affecting their location either as offset to the attachment or general position in world Misc: fixed small bugs

+++ code/libs/HalfEdgeMesh2D/HalfEdgeMesh2D.cpp
@@ -60,7 +60,7 @@ void HalfEdgeMesh2D::Construct(const char * path)
 		map += str;
 		height++;
 	}
-	width = str.length();
+	width = (int)str.length();
 	Vector<Face*> allFaces;
 	allFaces.reserve(height*width);
 	Face* emptyFace = facePool.PoolPartyAlloc();
