+++ code/libs/HalfEdgeMesh2DSquared/HalfEdgeMesh2D.cpp
@@ -34,7 +34,7 @@ void HalfEdgeMesh2D::Construct(const char * path)
 		map += str;
 		height++;
 	}
-	width = str.length();
+	width = (int)str.length();
 	Vector<Face*> allFaces;
 	allFaces.reserve(height*width);
 	Face* emptyFace = facePool.PoolPartyAlloc();
